Fixed git contension
