+++ tests/conftest.py
@@ -1,5 +1,7 @@
 import pytest
 
+from cardex.backend.backend_base import AbstractBackend
+from cardex.backend.dbsync import DbsyncBackend
 from cardex.dexs.core.base import AbstractPairState
 
 # This grabs all the DEXs
@@ -26,6 +28,18 @@ DEXS = [pytest.param(d, marks=getattr(pytest.mark, d.dex.lower())) for d in D]
 
 @pytest.fixture(scope="module", params=DEXS)
 def dex(request) -> AbstractPairState:
+    """Autogenerate a list of all DEX classes.
+
+    Returns:
+        List of all DEX classes. This could be a full order book, an individual order,
+        a stableswap, or a constant product pool class.
+    """
+
+    return request.param
+
+
+@pytest.fixture(scope="module", params=[DbsyncBackend()])
+def backend(request) -> AbstractBackend:
     """Autogenerate a list of all DEX classes.
 
     Returns:
