+++ tests/test_amm.py
@@ -5,6 +5,7 @@ import pytest
 from cardex import MinswapDJEDiUSDStableState
 from cardex import MinswapDJEDUSDCStableState
 from cardex import MinswapDJEDUSDMStableState
+from cardex import MinswapV2CPPState
 from cardex import SundaeSwapV3CPPState
 from cardex import WingRidersSSPState
 from cardex.dexs.amm.amm_base import AbstractPoolState
@@ -25,34 +26,38 @@ MALFORMED_CBOR = {
 }
 
 
-def test_pools_script_version(dex: AbstractPoolState, subtests):
-    if issubclass(dex, AbstractOrderBookState):
-        return
-
-    selector = dex.pool_selector
-    result = AbstractPoolState.get_backend().get_pool_utxos(
-        limit=1, historical=False, **selector.to_dict()
-    )
-
-    counts = 0
-    for pool in result:
-        try:
-            dex.model_validate(pool.model_dump())
-            counts += 1
-        except (InvalidLPError, NoAssetsError, InvalidPoolError):
-            pass
-        except:
-            raise
-
-
-def test_parse_pools(dex: AbstractPoolState, run_slow: bool, subtests):
+def test_get_pool_script_version(dex: AbstractPoolState, benchmark, backend):
     if issubclass(dex, AbstractOrderBookState):
         return
 
     selector = dex.pool_selector()
+    result = benchmark(
+        backend.get_pool_utxos,
+        limit=1,
+        historical=False,
+        **selector.model_dump(),
+    )
+    if dex.dex in ["Spectrum"] or dex in [
+        MinswapDJEDiUSDStableState,
+        MinswapDJEDUSDCStableState,
+        MinswapDJEDUSDMStableState,
+        SundaeSwapV3CPPState,
+        MinswapV2CPPState,
+    ]:
+        assert result[0].plutus_v2
+    else:
+        assert not result[0].plutus_v2
+
+
+def test_parse_pools(dex: AbstractPoolState, run_slow: bool, subtests, backend):
+    if issubclass(dex, AbstractOrderBookState):
+        return
+
+    AbstractPoolState.set_backend(backend)
+    selector = dex.pool_selector()
     limit = 20000 if run_slow else 100
     result = AbstractPoolState.get_backend().get_pool_utxos(
-        limit=limit, historical=False, **selector.to_dict()
+        limit=limit, historical=False, **selector.model_dump()
     )
 
     counts = 0
